# root/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/wastage-reports": {
      "post": {
        "operationId": "WastageReportsController_create",
        "summary": "Create Wastage Report",
        "description": "Submits a new wastage report. Any authenticated user (including Campus Visitor) can report wastage. Send a POST with a CreateWastageReportDto JSON body.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWastageReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Wastage report created successfully."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          }
        },
        "tags": [
          "wastage-reports"
        ]
      },
      "get": {
        "operationId": "WastageReportsController_findAll",
        "summary": "List All Wastage Reports",
        "description": "Retrieves all wastage reports. Any authenticated user can view them (the frontend filters by reporter).",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of wastage report records returned."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          }
        },
        "tags": [
          "wastage-reports"
        ]
      }
    },
    "/api/wastage-reports/{id}/photos": {
      "post": {
        "operationId": "WastageReportsController_uploadPhotos",
        "summary": "Attach Wastage Report Photos",
        "description": "Uploads up to 4 photos as evidence for a wastage report. Any authenticated user can attach photos. Send a multipart/form-data POST request with images under the 'files' field.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Photos attached successfully."
          },
          "400": {
            "description": "File missing, wrong type, or over the size limit."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Wastage report with the given ID not found."
          }
        },
        "tags": [
          "wastage-reports"
        ]
      }
    },
    "/api/wastage-reports/{id}": {
      "get": {
        "operationId": "WastageReportsController_findOne",
        "summary": "Get Wastage Report by ID",
        "description": "Retrieves a single wastage report by UUID. Any authenticated user can look up reports.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wastage report record returned."
          },
          "403": {
            "description": "Forbidden (RBAC)"
          },
          "404": {
            "description": "Not found."
          }
        },
        "tags": [
          "wastage-reports"
        ]
      },
      "put": {
        "operationId": "WastageReportsController_put",
        "summary": "Replace Wastage Report",
        "description": "Completely replaces an existing wastage report. Only the Organization Admin can perform full replacements.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutWastageReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced successfully."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Not found."
          }
        },
        "tags": [
          "wastage-reports"
        ]
      },
      "patch": {
        "operationId": "WastageReportsController_update",
        "summary": "Update Wastage Report",
        "description": "Partially updates a wastage report (e.g., status). The Sustainability Officer and Organization Admin can update reports.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWastageReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated successfully."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Not found."
          }
        },
        "tags": [
          "wastage-reports"
        ]
      },
      "delete": {
        "operationId": "WastageReportsController_remove",
        "summary": "Delete Wastage Report",
        "description": "Permanently removes a wastage report. Only the Organization Admin can delete.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "User role for RBAC.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Not found."
          }
        },
        "tags": [
          "wastage-reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "EnerTrack API",
    "description": "Backend API for EnerTrack — Facility Management, Sustainability & Finance Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      },
      "x-org-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-org-id"
      }
    },
    "schemas": {
...
      "CreateWastageReportDto": {
        "type": "object",
        "properties": {},
        "example": {
          "organization_id": "org-001",
          "reporter_id": "550e8400-0002-4000-8000-000000000002",
          "type": "Water",
          "status": "Forwarded to Finance",
          "details": {
            "specificData": {
              "nature": "Sprinkler system leaking heavily near the pathway.",
              "location": "Main Courtyard, East Wing"
            },
            "priority": "High",
            "systemData": {
              "sensorId": "FLOW-CW-01",
              "readingValue": 45.5,
              "readingUnit": "L/min",
              "baselineValue": 20,
              "confidence": "High",
              "status": "abnormal"
            }
          },
          "sensor_reading_id": "rrrr0000-0001-4000-8000-000000000000"
        }
      },
      "PutWastageReportDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWastageReportDto": {
        "type": "object",
        "properties": {},
        "example": {
          "organization_id": "org-001",
          "reporter_id": "550e8400-0002-4000-8000-000000000002",
          "type": "Water",
          "status": "Forwarded to Finance",
          "details": {
            "specificData": {
              "nature": "Sprinkler system leaking heavily near the pathway.",
              "location": "Main Courtyard, East Wing"
            },
            "priority": "High",
            "systemData": {
              "sensorId": "FLOW-CW-01",
              "readingValue": 45.5,
              "readingUnit": "L/min",
              "baselineValue": 20,
              "confidence": "High",
              "status": "abnormal"
            }
          },
          "sensor_reading_id": "rrrr0000-0001-4000-8000-000000000000"
        }
      },
...
    }
  }
}